refactor: Convert PageControl to generic type PageControl<TObservable>
TObservable must inherit from ObservableObject.

Add Source property to allow the derived type to have the DataSource automatically adjusted if the DataContext changes.

When the control is loaded, it will reach out to the application's active service provider to retrieve the view model for the control.

Add dummy types that allow the designer to load generic types that inherit from UserControl.

Renamed SettingsViews to SettingsView.

Add a binding source object to every page control.