refactor: Implement LaunchInterceptableClientAsync
Prior to launching any game client, we need to attempt to load up a .json file that contains the barebones required game information that relates to a specifics client version(MD5).

Any HConnection instances created will be stored in a dictionary, and have the ticket that was used to initialize the session as their key.